Add synset adding to java model.

+++ src/main/java/com/github/semres/gui/DatabasesController.java
@@ -1,5 +1,6 @@
 package com.github.semres.gui;
 
+import com.github.semres.Board;
 import com.github.semres.SemRes;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
@@ -10,15 +11,17 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.ListView;
+import javafx.scene.input.MouseEvent;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
 
 import java.io.IOException;
+import java.lang.reflect.InvocationTargetException;
 import java.net.URL;
 import java.util.ResourceBundle;
 import java.util.Set;
 
-public class DatabasesController extends Controller implements Initializable {
+public class DatabasesController extends ChildController implements Initializable {
     @FXML
     private ListView listView;
     @FXML
@@ -60,9 +63,19 @@ public class DatabasesController extends Controller implements Initializable {
 
     public void deleteDatabase() {
         String name = (String)listView.getSelectionModel().getSelectedItem();
-        if (name != null) {
+        if (name != null && !name.equals("SYSTEM")) {
             SemRes.getInstance().deleteRepository(name);
             observableList.remove(name);
+        }
+    }
+
+    public void clickedListView(MouseEvent click) throws NoSuchMethodException, InstantiationException, IllegalAccessException, InvocationTargetException {
+        if (click.getClickCount() == 2 && listView.getSelectionModel().getSelectedItem() != null) {
+            String repositoryName = (String) listView.getSelectionModel().getSelectedItem();
+            Board loadedBoard =  SemRes.getInstance().getBoard(repositoryName);
+            ((MainController) parent).setBoard(loadedBoard);
+            Stage stage = (Stage) listView.getScene().getWindow();
+            stage.close();
         }
     }
 }
